requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev.in
#
alabaster==0.7.12         # via sphinx
appnope==0.1.0            # via ipython
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
babel==2.7.0              # via sphinx
backcall==0.1.0           # via ipython
certifi==2019.6.16
chardet==3.0.4
click==7.0                # via safety
cloudinary==1.16.0
coverage==4.5.3           # via django-coverage-plugin, pytest-cov
decorator==4.4.0          # via ipython, traitlets
django-coverage-plugin==1.6.0
django-debug-toolbar==1.11
django-embed-video==1.2.0
django-environ==0.4.5
django-extensions==2.1.9
django==1.11.22
docutils==0.14            # via sphinx
dparse==0.4.1             # via safety
entrypoints==0.3          # via flake8
factory-boy==2.12.0
faker==1.0.7              # via factory-boy
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0          # via sphinx
importlib-metadata==0.18  # via pluggy, pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.6.1
jedi==0.14.0              # via ipython
jinja2==2.10.1            # via sphinx
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mock==3.0.5
more-itertools==7.1.0     # via pytest
multidict==4.5.2          # via yarl
packaging==19.0           # via dparse, pytest, safety, sphinx
parso==0.5.0              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.12.0            # via pytest
prompt-toolkit==2.0.9     # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pygments==2.4.2           # via ipython, sphinx
pyparsing==2.4.0          # via packaging
pytest-cov==2.7.1
pytest-django==3.5.1
pytest==5.0.1             # via pytest-cov, pytest-django
python-dateutil==2.8.0    # via faker
pytz==2019.1
pyyaml==5.1.1
requests==2.22.0
safety==1.8.5
selenium==3.141.0
six==1.12.0
snowballstemmer==1.9.0    # via sphinx
sphinx-rtd-theme==0.4.3
sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1  # via sphinx
sphinxcontrib-devhelp==1.0.1  # via sphinx
sphinxcontrib-htmlhelp==1.0.2  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.2  # via sphinx
sphinxcontrib-serializinghtml==1.1.3  # via sphinx
sqlparse==0.3.0           # via django-debug-toolbar
text-unidecode==1.2       # via faker
traitlets==4.3.2          # via ipython
urllib3==1.25.3
vcrpy==2.0.1
wcwidth==0.1.7            # via prompt-toolkit, pytest
werkzeug==0.15.4
whitenoise==3.3.1
wrapt==1.11.2             # via vcrpy
yarl==1.3.0               # via vcrpy
zipp==0.5.1               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via ipython, safety, sphinx
